16263: Adds test exposing a bug when using 'limit' with the max int64 value.
Arvados-DCO-1.1-Signed-off-by: Lucas Di Pentima <lucas@di-pentima.com.ar>

+++ sdk/go/arvados/api.go
@@ -77,8 +77,8 @@ type ListOptions struct {
 	Select             []string               `json:"select"`
 	Filters            []Filter               `json:"filters"`
 	Where              map[string]interface{} `json:"where"`
-	Limit              int                    `json:"limit"`
-	Offset             int                    `json:"offset"`
+	Limit              int64                  `json:"limit"`
+	Offset             int64                  `json:"offset"`
 	Order              []string               `json:"order"`
 	Distinct           bool                   `json:"distinct"`
 	Count              string                 `json:"count"`
